Refactoring Domain and restructuring.
