+++ calimap_server/app.py
@@ -1,0 +1,17 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+from domain.repository import Repository
+from application.api.v1 import alphabet
+from ui import controller
+
+app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db/calimap.sqlite'
+
+Repository.db = SQLAlchemy(app)
+
+app.register_blueprint(alphabet.blueprint, url_prefix='/api/v1')
+app.register_blueprint(controller.blueprint, url_prefix='/')
+
+if __name__ == '__main__':
+    app.run(debug=True, port=5000)
