+++ calimap_server/domain/config.py
@@ -1,0 +1,2 @@
+class Config:
+    dbPath = "sqlite:///db/calimap.sqlite'"
