+++ calimap_server/domain/repository.py
@@ -1,0 +1,8 @@
+from flask_sqlalchemy import SQLAlchemy
+
+
+class Repository:
+    def __init__(self):
+        pass
+
+    db = SQLAlchemy
