+++ calimap_server/requirements/requirements.txt
@@ -1,0 +1,3 @@
+Flask==0.11
+Flask_SQLAlchemy==2.0
+SQLAlchemy==0.9.8
